Revert "Configured project to build a library called CardGame."
This reverts commit b62f5080d925e888bc1672e31bf14161db6522f4.

Reverting commit to simplify development of game. The War game
implementation will live in the project simply as another folder. In the
future the main CardGame logic can be turned into a library, but to keep
the development as simple as possible for now, it will stay in the same
project as the War Game implementation.